requirements.txt
pip~=25.0.1
Django~=5.1.7
asgiref~=3.8.1
sqlparse~=0.5.3
psycopg2~=2.9.10
djangorestframework~=3.15.2
python-dotenv~=1.0.1